Framework of an MCP/2.1 implementation
Currently it finds MCP messages in the output stream, parses them,
and will dispatch them to registered packages.  It creates and honors an
auth key.

The 'mcp' 'mcp-negotiation' and 'mcp-cord' implementations are not at
all complete.  The registry scheme is cart-before-horse hackish.
There's no version checking for packages.

Basically it knows just enough to shake hands with an MCP/2.1 server and
exchange mcp-negotiate information (which it discards without reading).

A few new_ok tests.

Move ANSI and MCP21 out of the misleadingly-named "Plugins" directory.