Multiline MCP messages working, to a first approximation
Got multiline doing the right thing, at least for the purposes of
understanding the multilines sent by dns-org-mud-moo-simpleedit

And, therefore, a stubbed-out package for dns-org-mud-moo-simpleedit

Starting to look like the 'mcp' package should just get absorbed back
into the core file.

And the multiline logic is *painful*.  Some love to be had there.