MCP 2.1 negotiation of versioning now working
It's still a little fragile:

Packages need to register with min and max versions, which are strings
of the form '#.#' -- this is mandatory but not ever checked for.

TODO: make it DWIM if you just say "version => 1" ?

Packages now will not receive dispatched messages until they're
'activated' by having seen an appropriate mcp-negotiate-can from the
server.

Also de-spam STDERR a little.