Step away from perl5i
The dependency list for perl5i is super long, and mostly I was using it
for auto-strict and -warnings, and the 'sub' and 'func' keywords.  I
tried dragging in Method::Signatures to do that but it dragged in all
of PPI which no thank you.

This commit retargets this whole thing at perl 5.14, with as few modules
as makes sense, since I intend to bundle this up into binaries someday.