Get rid of id(); a few style twiddles
So now that I grok wxperl a little better, this whole C++-style
id(WINDOW_BIG_CONSTANT) business is dumb in this context.  It's gone
now, and we use actual perl variables because perl.

Also some style twiddles to the main frame;  and add some accessors for
things to find each other more easily (ie, output and input panes)

Now the 'worlds' data structure will get populated if it's in the prefs
file.  Someday we'll even display it in the dialog.

Remove the t/ file that was testing the now-gone id() sub.