Not-finished implementation of X-Windows copypaste
This is still broken in that EVT_RICHTEXT_SELECTION_CHANGED seems to be
exported by Wx::RichText but isn't actually a "macro" or something, so
it simply breaks if it's used.

Middle-mouse click-to-paste and Shift-Insert are both working fine now.

This also adds a use_x_copy_paste preference, but the whole defaults
mechanism seems broken so it has to be added into .WxMOO manually.